freqfusion_fasterrcnn/requirements.txt
torch>=1.10
torchvision>=0.11
pycocotools>=2.0.6
numpy>=1.21
pyyaml>=6.0
tqdm>=4.62
Pillow>=9.0
